modules/16bit/code/16bit_e2e/requirements.txt
pyserial
mido
sounddevice
soundfile
scipy 